fixed data type discrepancy in comparison

+++ tasks/rozenberg_a_matrix_column_sum/tests/functional/main.cpp
@@ -39,14 +39,14 @@ class RozenbergAMatrixColumnFuncTests : public ppc::util::BaseRunFuncTests<InTyp
       file >> rows >> columns;
 
       InType input_data(rows, std::vector<int>(columns));
-      for (size_t i = 0; i < rows; i++) {
-        for (size_t j = 0; j < columns; j++) {
+      for (int i = 0; i < rows; i++) {
+        for (int j = 0; j < columns; j++) {
           file >> input_data[i][j];
         }
       }
 
       OutType output_data(columns);
-      for (size_t i = 0; i < columns; i++) {
+      for (int i = 0; i < columns; i++) {
         file >> output_data[i];
       }
       input_data_ = input_data;
